Establish `cell` directory and move all related code there
-  cellClass, cellRenderer, onCellClick and cellUtils files created
-  code previously located in other files split off and moved

JIRA: FET-608
